OSRA-326 Partner Facade
DO NOT MERGE

This is my idea of what a facade should do: package data for presentation to
the view. In the partners controller, its utility is negligible since most
actions do not instantiate multiple variables and their interface to the views
is already minimal. With that in mind, having explored the use of a facade in
the partners controller as per our initial intent, I am not certain that its
benefits outweigh the costs. Other controllers, however, may well benefit from
facades. Sponsor show & User show both display the main object's associations,
and these associations feel like ideal candidates for loading from a facade.

In PR278, the facade takes on multiple responsibilities and thus violates
SOLID guidelines. In addition to sending data from the controller to the view,
it also gets involved in building and updating the partner object with
attributes returned from the view to the controller. The facade also takes
over validation of params, which in my mind clearly falls into the controller's
area of responsibility.